requirements.txt
numpy
scipy
pillow
ipykernel